+++ app/src/main/java/afkt/project/database/green/GreenManager.java
@@ -5,7 +5,8 @@ import android.text.TextUtils;
 import java.util.HashMap;
 import java.util.Map;
 
-import afkt.project.database.green.able.GreenDatabase;
+import afkt.project.database.green.able.AbsGreenDatabase;
+import afkt.project.database.green.module.image.ImageDatabase;
 import afkt.project.database.green.module.note.NoteDatabase;
 import dev.utils.app.logger.DevLogger;
 import dev.utils.common.StringUtils;
@@ -13,6 +14,16 @@ import dev.utils.common.StringUtils;
 /**
  * detail: GreenDao 管理类
  * @author Ttt
+ * <pre>
+ *     官方文档
+ *     @see <a href="https://greenrobot.org/greendao/documentation/modelling-entities"/>
+ *     SQL 语句写到累了? 试试 GreenDAO
+ *     @see <a href="https://www.jianshu.com/p/11bdd9d761e6"/>
+ *     Android GreenDao 数据库
+ *     @see <a href="https://www.jianshu.com/p/26c60d59e76d"/>
+ *     Android ORM 框架 : GreenDao 使用详解 ( 进阶篇 )
+ *     @see <a href="https://blog.csdn.net/speedystone/article/details/74193053"/>
+ * </pre>
  */
 public final class GreenManager {
 
@@ -22,8 +33,8 @@ public final class GreenManager {
     // 日志 TAG
     private static final String TAG = GreenManager.class.getSimpleName();
 
-    // DataBase 对象缓存
-    private static final Map<String, GreenDatabase> sDatabaseMaps = new HashMap<>();
+    // Database 对象缓存
+    private static final Map<String, AbsGreenDatabase> sDatabaseMaps = new HashMap<>();
 
     // ============
     // = database =
@@ -32,10 +43,10 @@ public final class GreenManager {
     /**
      * 获取 GreenDatabase 对象
      * @param dbName 数据库名
-     * @param clazz  {@link GreenDatabase} 实现类
-     * @return {@link GreenDatabase}
+     * @param clazz  {@link AbsGreenDatabase} 实现类
+     * @return {@link AbsGreenDatabase}
      */
-    public static <T extends GreenDatabase> T database(final String dbName, final Class clazz) {
+    public static <T extends AbsGreenDatabase> T database(final String dbName, final Class clazz) {
         return database(dbName, null, clazz);
     }
 
@@ -43,11 +54,11 @@ public final class GreenManager {
      * 获取 GreenDatabase 对象
      * @param dbName   数据库名
      * @param password 数据库解密密码
-     * @param clazz    {@link GreenDatabase} 实现类
-     * @return {@link GreenDatabase}
+     * @param clazz    {@link AbsGreenDatabase} 实现类
+     * @return {@link AbsGreenDatabase}
      */
-    public static <T extends GreenDatabase> T database(final String dbName, final String password,
-                                                       final Class clazz) {
+    public static <T extends AbsGreenDatabase> T database(final String dbName, final String password,
+                                                          final Class clazz) {
         if (TextUtils.isEmpty(dbName)) return null;
 
         // 获取数据库名
@@ -60,7 +71,7 @@ public final class GreenManager {
                 DevLogger.eTag(TAG, e, "database");
             }
         }
-        GreenDatabase greenDatabase = sDatabaseMaps.get(databaseName);
+        AbsGreenDatabase greenDatabase = sDatabaseMaps.get(databaseName);
         if (greenDatabase != null) {
             T db = null;
             try {
@@ -78,26 +89,32 @@ public final class GreenManager {
     // ========
 
     // 数据库创建接口
-    private static final GreenDatabase.Create CREATE = new GreenDatabase.Create() {
+    private static final AbsGreenDatabase.Create CREATE = new AbsGreenDatabase.Create() {
 
         @Override
         public String getDatabaseName(String dbName, String password, Class clazz) {
-            return GreenDatabase.createDatabaseName(dbName, StringUtils.isNotEmpty(password));
+            return AbsGreenDatabase.createDatabaseName(dbName, StringUtils.isNotEmpty(password));
         }
 
         @Override
-        public GreenDatabase create(String dbName, String password, Class clazz) {
+        public AbsGreenDatabase create(String dbName, String password, Class clazz) {
             if (clazz == NoteDatabase.class) {
                 return NoteDatabase.database(dbName, password);
+            } else if (clazz == ImageDatabase.class) {
+                return ImageDatabase.database(dbName, password);
             }
             return null;
         }
     };
 
-    // ===============
-    // = 快捷获取方法 =
-    // ===============
+    // ============
+    // = 快捷方法 =
+    // ============
 
+    /**
+     * 获取 Note Database
+     * @return {@link NoteDatabase}
+     */
     public static NoteDatabase getNoteDatabase() {
         return database(NoteDatabase.TAG, NoteDatabase.class);
     }