+++ app/src/main/java/afkt/project/database/room/able/AbsRoomDatabase.java
@@ -1,21 +1,18 @@
-package afkt.project.database.green.able;
+package afkt.project.database.room.able;
 
-import org.greenrobot.greendao.database.Database;
-
-import gen.greendao.DaoMaster;
-import gen.greendao.DaoSession;
+import androidx.room.RoomDatabase;
 
 /**
- * detail: Green Database 抽象类
+ * detail: Room Database 抽象类
  * @author Ttt
  * <pre>
  *     dbName 最好为 功能模块名, 如果涉及需要区分用户则为 功能模块名 + 用户 id
  * </pre>
  */
-public abstract class GreenDatabase {
+public abstract class AbsRoomDatabase extends RoomDatabase {
 
     // 数据库名
-    private static final String DATABASE_NAME = "green-db";
+    private static final String DATABASE_NAME = "room-db";
 
     /**
      * 拼接数据库名
@@ -27,32 +24,6 @@ public abstract class GreenDatabase {
         if (encrypted) return dbName + "-" + DATABASE_NAME + "-encrypted";
         return dbName + "-" + DATABASE_NAME;
     }
-
-    // =
-
-    /**
-     * 获取 DaoMaster.OpenHelper
-     * @return {@link DaoMaster.OpenHelper}
-     */
-    public abstract DaoMaster.OpenHelper getHelper();
-
-    /**
-     * 获取 Database
-     * @return {@link Database}
-     */
-    public abstract Database getDatabase();
-
-    /**
-     * 获取 DaoMaster
-     * @return {@link DaoMaster}
-     */
-    public abstract DaoMaster getDaoMaster();
-
-    /**
-     * 获取 DaoSession
-     * @return {@link DaoSession}
-     */
-    public abstract DaoSession getDaoSession();
 
     // =============
     // = 创建数据库 =
@@ -68,7 +39,7 @@ public abstract class GreenDatabase {
          * 获取数据库名
          * @param dbName   数据库名
          * @param password 数据库解密密码
-         * @param clazz    {@link GreenDatabase} 实现类
+         * @param clazz    {@link AbsRoomDatabase} 实现类
          * @return 数据库名
          */
         String getDatabaseName(String dbName, String password, Class clazz);
@@ -77,9 +48,9 @@ public abstract class GreenDatabase {
          * 创建数据库方法
          * @param dbName   数据库名
          * @param password 数据库解密密码
-         * @param clazz    {@link GreenDatabase} 实现类
-         * @return {@link GreenDatabase}
+         * @param clazz    {@link AbsRoomDatabase} 实现类
+         * @return {@link AbsRoomDatabase}
          */
-        GreenDatabase create(String dbName, String password, Class clazz);
+        AbsRoomDatabase create(String dbName, String password, Class clazz);
     }
 }