+++ lib/DevApp/src/main/java/dev/utils/app/helper/DevHelper.java
@@ -22,6 +22,9 @@ import androidx.fragment.app.DialogFragment;
 
 import java.io.Closeable;
 import java.io.File;
+import java.io.Flushable;
+import java.io.OutputStream;
+import java.io.Writer;
 import java.util.Locale;
 
 import dev.utils.app.AnalysisRecordUtils;
@@ -1329,6 +1332,66 @@ public final class DevHelper {
      */
     public DevHelper closeIOQuietly(final Closeable... closeables) {
         CloseUtils.closeIOQuietly(closeables);
+        return this;
+    }
+
+    /**
+     * 将缓冲区数据输出
+     * @param flushables Flushable[]
+     * @return {@link DevHelper}
+     */
+    public DevHelper flush(final Flushable... flushables) {
+        CloseUtils.flush(flushables);
+        return this;
+    }
+
+    /**
+     * 安静将缓冲区数据输出
+     * @param flushables Flushable[]
+     * @return {@link DevHelper}
+     */
+    public DevHelper flushQuietly(final Flushable... flushables) {
+        CloseUtils.flushQuietly(flushables);
+        return this;
+    }
+
+    /**
+     * 将缓冲区数据输出并关闭流
+     * @param outputStream {@link OutputStream}
+     * @return {@link DevHelper}
+     */
+    public DevHelper flushCloseIO(final OutputStream outputStream) {
+        CloseUtils.flushCloseIO(outputStream);
+        return this;
+    }
+
+    /**
+     * 安静将缓冲区数据输出并关闭流
+     * @param outputStream {@link OutputStream}
+     * @return {@link DevHelper}
+     */
+    public DevHelper flushCloseIOQuietly(final OutputStream outputStream) {
+        CloseUtils.flushCloseIOQuietly(outputStream);
+        return this;
+    }
+
+    /**
+     * 将缓冲区数据输出并关闭流
+     * @param writer {@link Writer}
+     * @return {@link DevHelper}
+     */
+    public DevHelper flushCloseIO(final Writer writer) {
+        CloseUtils.flushCloseIO(writer);
+        return this;
+    }
+
+    /**
+     * 安静将缓冲区数据输出并关闭流
+     * @param writer {@link Writer}
+     * @return {@link DevHelper}
+     */
+    public DevHelper flushCloseIOQuietly(final Writer writer) {
+        CloseUtils.flushCloseIOQuietly(writer);
         return this;
     }